+++ include/SpaceInvaders/IoController.h
@@ -25,9 +25,7 @@ SOFTWARE.
 
 #include <array>
 #include <atomic>
-#include <bitset>
 #include <memory>
-#include <mutex>
 
 #include "Base/Base.h"
 #include "SpaceInvaders/MemoryController.h"
@@ -135,8 +133,6 @@ namespace SpaceInvaders
 			nullptr							/**< Unused */
 		};
 
-		std::mutex mutex_;
-		std::array<uint8_t, MemoryController::GetVramLength()> vram_;
 	public:
 		/** Initialisation contructor.
 
@@ -228,9 +224,9 @@ namespace SpaceInvaders
 			@param	port		The output device to write to.
 			@param	data		The data to write to the output device.
 
-			@return	uint16_t	1 if the data on the port was handled, 0 otherwise.
+			@return				Audio that requires rendering as described above.
 		*/
-		std::bitset<16> WriteTo(uint16_t port, uint8_t data);
+		uint8_t WriteTo(uint16_t port, uint8_t data);
 
 		/** Service io interrupts.
 
@@ -249,11 +245,12 @@ namespace SpaceInvaders
 			rotated a further 270 degrees so it can be rendered with the correct
 			orientation.
 
-			@param	texture		The video memory to write to.
+			@param	dst			The video memory to write to (texture memory).
+			@param	src			The video ram to copy.
 			@param	rowBytes	The width of each scanline in bytes.
 
 		*/
-		void Blit(uint8_t* texture, uint8_t rowBytes);
+		void Blit(uint8_t* dst, uint8_t* src, uint8_t rowBytes);
 	};
 } // namespace SpaceInvaders